Move utils into rubberduke.utils
