+++ src/main/java/rubberduke/utils/Storage.java
@@ -1,4 +1,6 @@
-package rubberduke;
+package rubberduke.utils;
+
+import rubberduke.UserException;
 
 import java.io.File;
 import java.io.FileNotFoundException;
