+++ src/main/java/rubberduke/utils/Ui.java
@@ -1,4 +1,4 @@
-package rubberduke;
+package rubberduke.utils;
 
 import java.util.Scanner;
